requirement.txt
pillow==12.0.0
pytube2==15.0.17
pyperclip==1.11.0
customtkinter
